feat: add hint and title ocean search dialog
