+++ ocean-components/src/main/res/drawable/icon_search.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M14.4121,14.4121L20,20"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#B6B9CC"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M10,16C13.3137,16 16,13.3137 16,10C16,6.6863 13.3137,4 10,4C6.6863,4 4,6.6863 4,10C4,13.3137 6.6863,16 10,16Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#B6B9CC"
+      android:strokeLineCap="round"/>
+</vector>
