+++ ocean-components/src/main/res/layout/ocean_dialog.xml
@@ -20,18 +20,50 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <FrameLayout
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:layout_marginStart="@dimen/ocean_spacing_stack_xs"
+        android:layout_marginEnd="@dimen/ocean_spacing_stack_xs"
+        android:orientation="vertical"
         app:layout_behavior="@string/appbar_scrolling_view_behavior">
+
+        <com.google.android.material.textfield.TextInputLayout
+            android:id="@+id/textInputLayout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:boxBackgroundMode="none"
+            app:hintEnabled="false"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/label"
+            app:passwordToggleTint="@color/ocean_color_interface_light_deep">
+
+            <com.google.android.material.textfield.TextInputEditText
+                android:id="@+id/inputSearch"
+                android:layout_width="match_parent"
+                android:drawableStart="@drawable/icon_search"
+                android:drawablePadding="@dimen/ocean_spacing_stack_xs"
+                android:layout_height="wrap_content"
+                android:background="@drawable/ocean_input_text_field_states"
+                android:fontFamily="@font/font_family_base_regular"
+                android:padding="@dimen/ocean_spacing_inline_xs"
+                android:textColor="@color/ocean_color_interface_dark_deep"
+                android:textColorHint="@color/ocean_color_interface_light_deep"
+                android:textCursorDrawable="@drawable/ocean_cursor"
+                android:textSize="@dimen/ocean_font_size_xs" />
+        </com.google.android.material.textfield.TextInputLayout>
+
+        <Space
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/ocean_spacing_stack_xs"/>
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerViewItems"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            app:layout_anchor="@+id/frameLayout"
             app:layout_anchorGravity="center" />
 
-    </FrameLayout>
+    </LinearLayout>
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
