+++ ocean-components/src/main/res/layout/ocean_text_list_item_inverted.xml
@@ -53,10 +53,6 @@
             name="arrowVisible"
             type="Boolean" />
 
-        <variable
-            name="highlight"
-            type="Boolean" />
-
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -102,10 +98,10 @@
                     style="@style/Ocean.Description"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
+                    android:layout_marginBottom="@dimen/ocean_spacing_stack_xxxs"
                     android:text="@{title}"
                     app:ocean_text="@{title}"
                     app:ocean_text_format="@{formatTypeTitle}"
-                    android:layout_marginBottom="@dimen/ocean_spacing_stack_xxxs"
                     tools:text="Title" />
 
                 <Space
@@ -128,10 +124,10 @@
                 style="@style/Ocean.Paragraph"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@{subtitle}"
                 android:layout_marginBottom="@dimen/ocean_spacing_stack_xxxs"
+                android:fontFamily="@font/font_family_base_medium"
+                android:text="@{subtitle}"
                 android:textColor="@color/ocean_color_interface_dark_deep"
-                app:highlight="@{highlight}"
                 app:ocean_text="@{subtitle}"
                 app:ocean_text_format="@{formatTypeSubTitle}"
                 app:ocean_visible_or_gone="@{subtitle != null}"
@@ -142,9 +138,9 @@
                 style="@style/Ocean.Paragraph"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
+                android:fontFamily="@font/font_family_base_medium"
                 android:text="@{description}"
                 android:textColor="@color/ocean_color_interface_dark_deep"
-                app:highlight="@{highlight}"
                 app:ocean_text="@{description}"
                 app:ocean_text_format="@{formatTypeDescription}"
                 app:ocean_visible_or_gone="@{description != null}"
